pyproject.toml
[project]
name = "python-paymcp-server-demo"
version = "0.0.1"
description = "Minimal MCP server demo with paymcp (Stripe/Walleot) that generates images via OpenAI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.11.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "paymcp>=0.0.4",
    "openai>=1.97.1",
]
